Add flag for anon users
