+++ AntGameAPI/app.js
@@ -20,7 +20,6 @@ const _serviceController = require("./controller/ServiceController");
 const _publicController = require("./controller/PublicController");
 
 const TokenHandler = require("./auth/WebTokenHandler");
-const TokenRevokedHandler = require("./handler/TokenRevokedHandler");
 const { RejectNotAdmin, ServiceEndpointAuth } = require("./auth/AuthHelpers");
 const responseTime = require("response-time");
 const { GetIpAddress } = require("./helpers/IpHelper");
@@ -35,6 +34,7 @@ const {
   failedLoginLimiter,
   registrationLimiter,
 } = require("./auth/RateLimiters");
+const { JwtResultHandler, TokenVerifier } = require("./helpers/Middleware");
 
 const UnauthenticatedRoutes = [
   "/auth/login",
@@ -47,11 +47,6 @@ const UnauthenticatedRoutes = [
   "/health",
   "/time",
 ];
-
-const send401 = (res, message) => {
-  res.status(401);
-  res.send(message);
-};
 
 initializeScheduledTasks();
 SpacesService.initializeConnection();
@@ -77,52 +72,8 @@ app.use(
   jwt({ secret: TokenHandler.secret, algorithms: ["HS256"] }).unless({
     path: UnauthenticatedRoutes,
   }),
-  function (err, req, res, next) {
-    if (!err) {
-      next();
-    } else if (err.code === "credentials_required") {
-      send401(res, "JWT required");
-      Logger.logAuthEvent("api hit without JWT", { ip: GetIpAddress(req) });
-      return;
-    } else if (err.code === "invalid_token") {
-      send401(res, "Invalid JWT");
-      Logger.logAuthEvent("api hit with invalid JWT", { ip: GetIpAddress(req) });
-      return;
-    }
-    Logger.logError("App.jwtHandler", err);
-    res.status(401);
-    res.send("Unauthorized");
-  },
-  async function (req, res, next) {
-    if (!req.user) {
-      next();
-      return;
-    }
-
-    if (req.user.anon !== true) {
-      const userID = req.user.id;
-      const adminToken = req.user.admin === true;
-      const TokenIsValid = await TokenRevokedHandler.isTokenValid(userID, adminToken);
-      if (TokenIsValid === false) {
-        Logger.logAuthEvent("invalid token received", {
-          userID,
-          adminToken,
-          username: req.user.username,
-        });
-        res.sendStatus(401);
-        return;
-      }
-    }
-
-    if (req.user.admin !== true) {
-      const LoginsEnabled = await TokenRevokedHandler.AreLoginsEnabled();
-      if (!LoginsEnabled) {
-        res.sendStatus(401);
-        return;
-      }
-    }
-    next();
-  }
+  JwtResultHandler,
+  TokenVerifier
 );
 
 //#region Admin
