+++ AntGameUI/src/antGame/Auth/LoginPage/LoginPage.module.css
@@ -16,6 +16,8 @@
   align-items: center;
   justify-content: center;
   border-radius: 0.5em;
+  padding: 0.5em;
+  margin: 0.25em 0.5em 0.25em 0;
 }
 
 .skipButton {
